Create atlas_merge.common
